# docs/swagger.yaml
basePath: /api
definitions:
  models.ChangeRecord:
    properties:
      change_diff:
        type: string
      change_type:
        type: string
      detected_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      site_id:
        type: string
      snapshot_hash:
        type: string
    type: object
  models.FileContent:
    properties:
      change_record_id:
        type: string
      content:
        type: string
      content_hash:
        type: string
      created_at:
        type: string
      file_path:
        type: string
      id:
        type: string
      site_id:
        type: string
    type: object
  models.MonitoredSite:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      url:
        type: string
    type: object
  models.SiteHooksResponse:
    properties:
      config:
        additionalProperties: true
        type: object
      content:
        type: string
      type:
        type: string
    type: object
  models.UploadChangeRequest:
    properties:
      change_diff:
        type: string
      change_type:
        enum:
        - js
        - content
        type: string
      content:
        description: 原始内容（可选）
        type: string
      file_path:
        type: string
      site_name:
        type: string
      site_url:
        type: string
      snapshot_hash:
        type: string
    required:
    - change_diff
    - change_type
    - site_url
    - snapshot_hash
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: 用于接收监控工具上报的变化数据，并提供站点 Hook 脚本配置。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: 站点监控服务 API
  version: "1.0"
paths:
  /changes:
    get:
      description: 根据站点ID和变化类型查询历史变化
      parameters:
      - description: 站点ID
        in: query
        name: site_id
        type: string
      - description: 变化类型 (js 或 content)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ChangeRecord'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 查询变化记录
      tags:
      - 变化记录
  /content:
    get:
      description: 根据站点ID、文件路径或哈希值查询历史文件内容
      parameters:
      - description: 站点ID
        in: query
        name: site_id
        required: true
        type: string
      - description: 文件路径
        in: query
        name: file_path
        type: string
      - description: 内容哈希值(MD5)
        in: query
        name: hash
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FileContent'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取文件内容快照
      tags:
      - 内容快照
  /site-hooks:
    get:
      description: 返回指定站点所有启用的 Hook 脚本（用于客户端注入）
      parameters:
      - description: 站点完整URL
        in: query
        name: url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SiteHooksResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取站点 Hook 脚本
      tags:
      - Hook脚本
  /sites:
    get:
      description: 返回所有被监控的站点列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.MonitoredSite'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 获取所有站点
      tags:
      - 站点管理
  /upload:
    post:
      consumes:
      - application/json
      description: 监控工具上传检测到的JS或内容变化
      parameters:
      - description: 上传数据
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UploadChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 上传变化记录
      tags:
      - 变化记录
swagger: "2.0"